+++ implementation/site/context/mockups.js
@@ -3,61 +3,62 @@
 	app.context('Mockups', {
 		
 		className: 'wrapper-full container-fluid',
-		template: ' ',
+		template: '<div region="mockups"></div>',
 
-		mockups: [
+		data: {
+			mockups: [
 
-			//breadcrumb
-			{tpl: 'breadcrumb.html'},
+				//breadcrumb
+				{tpl: 'breadcrumb.html'},
 
-			//navbars
-			{tpl: 'nav-bar.html', className: 'navbar-default'},
-			{tpl: 'nav-bar.html', className: 'navbar-inverse'},
+				//navbars
+				{tpl: 'nav-bar.html', className: 'navbar-default'},
+				{tpl: 'nav-bar.html', className: 'navbar-inverse'},
 
-			//boxes
-			{tpl: 'boxes.html', onReady: function(){
-				if(Modernizr.chrome){
-					this.$el.find('[warning="chrome"]').removeClass('hidden');
-				}
-			}},
+				//boxes
+				{tpl: 'boxes.html', onReady: function(){
+					if(Modernizr.chrome){
+						this.$el.find('[warning="chrome"]').removeClass('hidden');
+					}
+				}},
 
-			//containers
-			{tpl: 'containers.html'},			
+				//containers
+				{tpl: 'containers.html'},			
 
-			//buttons
-			{tpl: 'buttons.html',},
+				//buttons
+				{tpl: 'buttons.html',},
 
-			//typography
-			{tpl: 'typography.html'},
+				//typography
+				{tpl: 'typography.html'},
 
-			//indicators (alert, lable, badge and progress bar)
-			{tpl: 'indicators.html'},
+				//indicators (alert, lable, badge and progress bar)
+				{tpl: 'indicators.html'},
 
-			//navigators
-			{tpl: 'navs.html'},
+				//navigators
+				{tpl: 'navs.html'},
 
-			//popups & dialogs
-			{tpl: 'dialogs.html', onReady: function(){
-				this.$el.find('[data-toggle="popover"]').popover();
-				this.$el.find('[data-toggle="tooltip"]').tooltip();
-			}},
+				//popups & dialogs
+				{tpl: 'dialogs.html', onReady: function(){
+					this.$el.find('[data-toggle="popover"]').popover();
+					this.$el.find('[data-toggle="tooltip"]').tooltip();
+				}},
 
-			//table
-			{tpl: 'table.html'},
+				//table
+				{tpl: 'table.html'},
 
-			//forms
-			{tpl: 'forms.html'},
+				//forms
+				{tpl: 'forms.html'},
 
-			//copyright of the mockup collections above
-			{tpl: 'copyright.html'}
+				//copyright of the mockup collections above
+				{tpl: 'copyright.html'}
 
-		],
+		]},
 
 		onReady: function(){
-			_.each(this.mockups, function(m){
-				//////////////////////////////////////////////////
-				///Absolute manually managed view life cycle...///
-				//////////////////////////////////////////////////
+			_.each(this.get('mockups'), function(m){
+				/////////////////////////////////////////////////////////////////////////
+				///Manually managed view life cycle..without this.show('region',...)..///
+				/////////////////////////////////////////////////////////////////////////
 				//1. create it
 				var view = app.view({
 					className: 'wrapper-full',
@@ -70,10 +71,10 @@
 					}
 				}, true);
 				//2. render and insert it into DOM
-				this.$el.append(view.render().el);
+				this.getRegion('mockups').$el.append(view.render().el);
 				//3. connect the view life-cycle event seq: --render--[[show]]--ready 
 				view.triggerMethod('show'); 
-				//(ref: /lib+-/marionette/view.js, we refined the seq and added the ready e though.)
+				//(ref: /lib+-/marionette/view.js, we refined the seq and added a ready e.)
 
 			}, this);
 		}