+++ CHANGELOG.md
@@ -29,8 +29,9 @@ Change Log
 25. 'context-switched' global coop event will now emit with additional metadata as second param.
 26. Added view.activate(group, filterFn) for programmatic (silent feedback) activations;
 27. Added app.config.defaultView as an alias to app.config.defaultContext;
-28. Added ctxConfig to app.navigate({path:..., ctxConfig:...});
-29. Aligned local/remote data ready event timing; (both now after this.paper ready)
+28. Added viewConfig to app.navigate({path:..., viewConfig:...}) as 2nd param for navigate-to;
+29. Aligned local/remote data ready event timing; (both after this.paper ready)
+30. Aligned navigate-to and ready event timing; (always be show-navito-ready for static, form, local/remote data views in a nav chain)
 
 working on ...